requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
cssselect2==0.8.0
ecdsa==0.19.1
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
freetype-py==2.5.1
greenlet==3.3.0
html5lib==1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
oscrypto==1.3.0
pillow==12.0.0
pyasn1==0.6.1
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.4.2
pyperclip==1.11.0
pyserial==3.5
python-bidi==0.6.7
python-jose==3.5.0
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.6
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
svglib==1.6.0
tinycss2==1.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.6.2
webencodings==0.5.1
Werkzeug==3.1.4
xhtml2pdf==0.2.17
